Refactor components, update mobile detection logic
