+++ components/Provider/PoolProvider.tsx
@@ -1,5 +1,4 @@
 import React, { useState, ReactNode, useEffect } from "react";
-import { POOL_CONFIG_MAP } from "constants/pools";
 import { PoolContext } from "context/PoolContext";
 import { PoolConfig } from "interfaces/pool";
 import { useAppData } from "context/AppDataContext";
@@ -13,14 +12,13 @@ export const PoolProvider: React.FC<PoolProviderProps> = ({ children }) => {
   const [poolConfig, setPoolConfig] = useState<PoolConfig | undefined>(
     undefined,
   );
-  const { chainId } = useAppData();
+  const { config } = useAppData();
 
   useEffect(() => {
-    if (chainId && poolName) {
-      const config = POOL_CONFIG_MAP[chainId]?.[poolName];
-      setPoolConfig(config);
+    if (config && poolName) {
+      setPoolConfig(config?.[poolName]);
     }
-  }, [chainId, poolName]);
+  }, [config, poolName]);
 
   const value = {
     poolName,
