+++ components/Provider/ReloadContextProvider.tsx
@@ -12,6 +12,7 @@ export const ReloadContextProvider: React.FC<ReloadContextProviderProps> = ({
 
   const reload = useCallback(() => {
     setReloadKey((reloadKey) => reloadKey + 1);
+    console.log("setReloadKey");
   }, []);
   return (
     <ReloadContext.Provider value={{ reload, reloadKey }}>
